added chan type, tok type, and more complete lexing.
The chan type is currently useless but hopes to emulate the channel functionality in go.

The tok type will be used as the medium for transport of tokens across the channel, from the lexer to the parser or some other thread running concurrently.

The lexer is now somewhat more complete, lexing strings of +|-|<|>| characters.